feat: split canisters api and services
