+++ frontend/svelte/src/lib/services/canisters.services.ts
@@ -1,12 +1,9 @@
 import type { Identity } from "@dfinity/agent";
 import { get } from "svelte/store";
-import { NNSDappCanister } from "../canisters/nns-dapp/nns-dapp.canister";
+import { queryCanisters } from "../api/canisters.api";
 import type { CanisterDetails } from "../canisters/nns-dapp/nns-dapp.types";
-import { OWN_CANISTER_ID } from "../constants/canister-ids.constants";
-import { identityServiceURL } from "../constants/identity.constants";
 import { canistersStore } from "../stores/canisters.store";
 import { i18n } from "../stores/i18n";
-import { createAgent } from "../utils/agent.utils";
 
 export const listCanisters = async ({
   clearBeforeQuery,
@@ -15,6 +12,11 @@ export const listCanisters = async ({
   clearBeforeQuery?: boolean;
   identity: Identity | null | undefined;
 }) => {
+  if (!identity) {
+    // TODO: https://dfinity.atlassian.net/browse/L2-346
+    throw new Error(get(i18n).error.missing_identity);
+  }
+
   if (clearBeforeQuery === true) {
     canistersStore.setCanisters([]);
   }
@@ -24,24 +26,4 @@ export const listCanisters = async ({
   });
 
   canistersStore.setCanisters(canisters);
-};
-
-const queryCanisters = async ({
-  identity,
-}: {
-  identity: Identity | null | undefined;
-}): Promise<CanisterDetails[]> => {
-  if (!identity) {
-    // TODO: https://dfinity.atlassian.net/browse/L2-346
-    throw new Error(get(i18n).error.missing_identity);
-  }
-
-  const agent = await createAgent({ identity, host: identityServiceURL });
-  // ACCOUNTS
-  const nnsDapp: NNSDappCanister = NNSDappCanister.create({
-    agent,
-    canisterId: OWN_CANISTER_ID,
-  });
-
-  return nnsDapp.getCanisters({ certified: true });
 };
