+++ frontend/svelte/src/tests/lib/services/accounts.services.spec.ts
@@ -5,7 +5,6 @@ import {
 } from "../../../lib/services/accounts.services";
 import { mockMainAccount } from "../../mocks/accounts.store.mock";
 import { mockIdentity } from "../../mocks/auth.store.mock";
-import {AccountNotFoundError} from '../../../lib/canisters/nns-dapp/nns-dapp.errors';
 
 describe("accounts-services", () => {
   const spyLoadAccounts = jest
@@ -37,8 +36,11 @@ describe("accounts-services", () => {
   });
 
   it("should not add subaccount if no identity", async () => {
-    const call = async () => await addSubAccount({ name: "test subaccount", identity: null });
+    const call = async () =>
+      await addSubAccount({ name: "test subaccount", identity: null });
 
-    await expect(call).rejects.toThrow(Error("No identity found to create subaccount"));
+    await expect(call).rejects.toThrow(
+      Error("No identity found to create subaccount")
+    );
   });
 });
